+++ examples/nextjs-releases/components/ReleaseItem.tsx
@@ -2,14 +2,15 @@ import { Release } from "@onsetio/browser";
 import styles from "./release-item.module.css";
 
 export default function ReleaseItem({ release }: { release: Release }) {
+  console.log("ReleaseItem", release);
   return (
     <article className={styles.release}>
       <h3 className={styles.title}>{release.title}</h3>
-      <div
+      {/* <div
         className={styles.description}
-        dangerouslySetInnerHTML={{ __html: release.description }}
-      ></div>
-      {release.change_list && (
+        dangerouslySetInnerHTML={{ __html: release.body }}
+      ></div> */}
+      {/* {release.change_list && (
         <ul className={styles.changes}>
           {release.change_list.map((change, index) => (
             <li
@@ -18,7 +19,7 @@ export default function ReleaseItem({ release }: { release: Release }) {
             ></li>
           ))}
         </ul>
-      )}
+      )} */}
     </article>
   );
 }