+++ packages/browser/src/types.ts
@@ -2,39 +2,22 @@ type Project = {
   id: string;
   name: string;
   slug: string;
+  url: string;
 };
 
 type Label = {
   id: string;
   name: string;
   color: string;
-};
-
-type LinkAttachment = {
-  type: "link";
-  content: {
-    url: string;
-    title: string;
-  };
-};
-
-type FileAttachment = {
-  type: "file";
-  content: {
-    url: string;
-    title: string;
-    size: number;
-    type: string;
-  };
+  slug: string;
+  url: string;
 };
 
 export interface Release {
   id: string;
   slug: string;
-  hero: {
-    url: string;
-    type: "image" | "video";
-  } | null;
+  hero_image: string | null;
+  hero_video: string | null;
   title: string;
   is_pinned: boolean;
   is_pre_release: boolean;
@@ -42,17 +25,16 @@ export interface Release {
   released_at: string;
   url: string;
   summary: string;
-  body: string;
-  body_text: string;
-  changes: {
+  description: string;
+  description_text: string;
+  change_list: {
     type: string;
     color: string;
-    content: string;
+    description: string;
     title: string;
   }[];
   labels: Label[];
   project: Project | null;
-  attachments: (LinkAttachment | FileAttachment)[];
 }
 
 export interface Milestone {
@@ -60,14 +42,30 @@ export interface Milestone {
   slug: string;
   title: string;
   votes: number;
+  rank: number;
   status: string;
-  stage: string;
+  stage: {
+    color: string;
+    title: string;
+  };
   url: string;
-  body: string;
-  body_text: string;
+  description: string;
+  description_text: string;
   project: Project | null;
   labels: Label[];
   created_at: string;
   updated_at: string;
-  attachments: (LinkAttachment | FileAttachment)[];
+  released_at: string | null;
+}
+
+export interface Workspace {
+  slug: string;
+  name: string;
+  color: string;
+  color_yiq: string;
+  url: string;
+  logo: string | null;
+  logo_dark: string | null;
+  projects: Omit<Project, "url">[];
+  labels: Omit<Label, "url">[];
 }